+++ src/constants.rs
@@ -1,6 +1,6 @@
 //! Constants.
 
-use trussed::types::{CertId, KeyId};
+use trussed_core::types::{CertId, KeyId};
 
 pub const FIDO2_UP_TIMEOUT: u32 = 30_000;
 pub const U2F_UP_TIMEOUT: u32 = 250;